requirements.txt
python-telegram-bot==20.7
gspread==6.0.0
oauth2client==4.1.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
APScheduler==3.10.4
python-dotenv==1.0.0
pytz==2024.1

